server/requirements.txt
Flask
gunicorn
pylint
psycopg2-binary